# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a GRPC Gateway API Documention server.",
        "title": "GRPC Gateway API Documention",
        "contact": {
            "name": "API Support",
            "url": "https://www.chantosweb.com",
            "email": "myrachanto1@gmail.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/logout": {
            "post": {
                "description": "Logout a new user item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Logout a user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/support.HttpError"
                        }
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "description": "GetAll a new user item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "GetAll a user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/users.UserDto"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/support.HttpError"
                        }
                    }
                }
            }
        },
        "/api/users/password": {
            "put": {
                "description": "PasswordUpdate a new user item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "PasswordUpdate a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User's oldpassword",
                        "name": "oldpassword",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User's newpassword",
                        "name": "newpassword",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/support.HttpError"
                        }
                    }
                }
            }
        },
        "/api/users/{code}": {
            "get": {
                "description": "GetOne a new user item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "GetOne a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/users.UserDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/support.HttpError"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a new user item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update a user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/support.HttpError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a new user item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Delete a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/support.HttpError"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "get the Health status of server.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health Status"
                ],
                "summary": "Show the Health status of server.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Login a new user item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Login a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User's Email",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User's Password",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/users.Auth"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/support.HttpError"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "Create a new user item",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Create a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User's First Name",
                        "name": "firstname",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User's Last Name",
                        "name": "lastname",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User's Phone Number",
                        "name": "phone",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User's Username",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User's Address",
                        "name": "address",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User's Email",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User's Password",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/users.UserDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/support.HttpError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "support.HttpError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "errors": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "results": {
                    "type": "boolean"
                }
            }
        },
        "users.Auth": {
            "type": "object",
            "properties": {
                "picture": {
                    "type": "string"
                },
                "reflesh_token": {
                    "type": "string"
                },
                "reflesh_token_expires": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "session_code": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "token_expires": {
                    "type": "string"
                },
                "usercode": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "users.Base": {
            "type": "object",
            "properties": {
                "created_At": {
                    "type": "string"
                },
                "delete_At": {
                    "type": "string"
                },
                "updated_At": {
                    "type": "string"
                }
            }
        },
        "users.UserDto": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "base": {
                    "$ref": "#/definitions/users.Base"
                },
                "birthday": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "picture": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "usercode": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}